# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses": {
      "get": {
        "operationId": "CoursesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/search": {
      "get": {
        "operationId": "CoursesController_searchAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchCourseDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}/skills/add": {
      "patch": {
        "operationId": "CoursesController_addSkills",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}/skills/remove": {
      "patch": {
        "operationId": "CoursesController_removeSkills",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDto"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/experiences": {
      "get": {
        "operationId": "ExperiencesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      },
      "post": {
        "operationId": "ExperiencesController_create",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperienceDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceDto"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      }
    },
    "/experiences/search": {
      "get": {
        "operationId": "ExperiencesController_searchAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchExperienceDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseDtoClass"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      }
    },
    "/experiences/{id}": {
      "get": {
        "operationId": "ExperiencesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceDto"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      },
      "patch": {
        "operationId": "ExperiencesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperienceDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceDto"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      },
      "delete": {
        "operationId": "ExperiencesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceDto"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      }
    },
    "/experiences/{id}/skills/add": {
      "patch": {
        "operationId": "ExperiencesController_addSkills",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperienceDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceDto"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      }
    },
    "/experiences/{id}/skills/remove": {
      "patch": {
        "operationId": "ExperiencesController_removeSkills",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperienceDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperienceDto"
                }
              }
            }
          }
        },
        "tags": [
          "experiences"
        ]
      }
    }
  },
  "info": {
    "title": "Profile Info",
    "description": "Api de informações pessoais feita em NestJS com Prisma",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationOutputDto": {
        "type": "object",
        "properties": {
          "totalRecords": {
            "type": "number"
          },
          "currentPage": {
            "type": "number"
          },
          "perPage": {
            "type": "number"
          },
          "nextPage": {
            "type": "number"
          },
          "prevPage": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "totalRecords",
          "currentPage",
          "perPage",
          "totalPages"
        ]
      },
      "PaginatedResponseDtoClass": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExperienceDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationOutputDto"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
...
      "CourseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "school": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ],
            "example": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ]
          },
          "certificate": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "school",
          "description",
          "status",
          "hours",
          "createdAt",
          "updatedAt",
          "userId"
        ]
      },
      "SearchCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "school": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ],
            "example": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ]
          },
          "certificate": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "school": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ],
            "example": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ]
          },
          "certificate": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "school",
          "description",
          "status",
          "hours"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "school": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ],
            "example": [
              "COMPLETE",
              "INCOMPLETE",
              "PAUSED",
              "PROGRESS"
            ]
          },
          "certificate": {
            "type": "string"
          },
          "hours": {
            "type": "number"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ExperienceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "organization",
          "description",
          "start",
          "createdAt",
          "updatedAt",
          "userId"
        ]
      },
      "SearchExperienceDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateExperienceDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "role",
          "organization",
          "description",
          "start"
        ]
      },
      "UpdateExperienceDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "start": {
            "type": "string"
          },
          "end": {
            "type": "string"
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}